Text change and refactor of about-section AboutIntro-component no longer a part of shared components - AboutIntro had alv-A specific text in it and does not belong under shared components. Removed unused duplicate aboutIntro-component from Alv-A - Removed tailwind-code-smells
closes #761
